prototype/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
sqlalchemy==2.0.35
psycopg[binary]==3.2.3
python-multipart==0.0.12
orjson==3.10.7
httpx==0.27.2
tenacity==9.0.0
python-dateutil==2.9.0.post0
